cfg: Fix count when creating new forwarder block
This was a bug previously but maybe did't matter as most of
the time the block was going to be removed after cddce.
Anyways the problem here is the count of the new bb was missing
of the first edge that was moved to it. So the fix is reuse the count
after the splitting of the edge as the initial value instead of 0.

This does not fix gcc.target/i386/pr90178.c with -m32, but at least
we don't get any more warning saying the incorrect count happening.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	PR tree-optimization/103680
	* tree-cfg.cc (make_forwarders_with_degenerate_phis): Fix
	initial value of the count to new bb.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>