ada: Reimplement AI12-0345, AI12-0372 and implement AI12-0402
The first two are binding interpretations, so apply to Ada 2012, and have
been only partially implemented.  The third is Ada 2022 and has not been
implemented, but is very convenient to tame the effects of the first two.

They mostly pertain to explicitly aliased parameters and to adjusting the
rules determining the master of a function call, so that the RM 6.4.1(6.4)
legality rule is applied judiciously.

The change also does some housekeeping work in the implementation of static
accessibility checks, plugging a few loopholes: in Ada 2005 for objects of
anonymous access types, in Ada 2005 and 2012 for conversions between access
types done in extended return statements, and in Ada 2012 for explicitly
aliased parameters.

gcc/ada/ChangeLog:

	* accessibility.ads (Is_Special_Aliased_Formal_Access): Delete.
	* accessibility.adb (Is_Special_Aliased_Formal_Access): Likewise.
	(Accessibility_Level.Innermost_Master_Scope_Depth): Look for non-
	package bodies and statements directly.
	(Accessibility_Level.Function_Call_Or_Allocator_Level): For a
	function call in a return context, return the extra level of the
	master of the call only for dynamic checks.
	(Accessibility_Level) <N_Defining_Identifier>: Always return the
	library level for an explicitly aliased parameter in the context
	of a return from the subprogram where it is declared.
	* exp_ch4.adb (Expand_N_Allocator): Test manually whether the
	context is a return statement instead of calling In_Return_Value.
	(Expand_N_Type_Conversion): Do not apply accessibility checks to
	actuals of a tagged type in a synthesized subprogram call.
	* sem_attr.adb (Resolve_Attribute) <Access>: Remove specific test
	for explicitly aliased parameters.
	* sem_ch4.adb: Remove clauses for Accessibility package.
	(Analyze_Call): Do not apply static accessibility checks here...
	* sem_res.adb (Resolve_Actuals): ...but here instead.  Implement
	a generalized form of AI12-0402.
	(Valid_Conversion): Apply static accessibility checks in extended
	return statements too.
	* sem_util.ads (In_Return_Value): Adjust description.
	(Is_Master): Delete.
	* sem_util.adb (Is_Explicitly_Aliased): Reindent.
	(In_Return_Value): Reimplement.
	(Is_Master): Delete.