MAINTAINERS file updates
So with the acquisition of Ventana Microsystems by Qualcomm, Daniel, Robin,
Shreya, Raphael and I all work for Qualcomm now and ongoing contributions will
be under that umbrella.

I also realized that Raphael has write-after-approval access, but didn't have a
suitable entry in the MAINTAINERS file.  So I updated that as well.

/

	* MAINTAINERS: Update entries for contributors formerly at
	Ventana Microsystems to refer to Qualcomm instead.