s390: Accept array arguments for builtins
A difference in front ends is that in contrast to C the C++ FE does no
array-to-pointer conversion prior calling resolve_overloaded_builtin().
However, we depend on this for finding the proper overloaded builtin or
in case of direct expansion.  Therefore, do the conversion manually.

gcc/ChangeLog:

	* config/s390/s390-c.cc (s390_resolve_overloaded_builtin):
	Perform array-to-pointer conversion for C++.

gcc/testsuite/ChangeLog:

	* g++.target/s390/builtin-array-arg-1.C: New test.
	* gcc.target/s390/builtin-array-arg-1.c: New test.