a68: support for -fmodules-map and -fmodules-map-file
This commit adds support for two new command-line options for the
Algol 68 front-end:

  -fmodules-map=<string>
  -fmodules-map-file=<filename>

These options are used in order to specify a mapping from module
indicants to file basenames.  The compiler will base its search for
the modules on these basenames rather on the default schema of
deriving the basename from the module indicant.

Signed-off-by: Jose E. Marchesi <jemarch@gnu.org>

gcc/algol68/ChangeLog

	* lang.opt (-fmodules-map): New option.
	(-fmodules-map-file): Likewise.
	* a68.h: Add prototype for a68_process_module_map.
	* a68-imports.cc (SKIP_WHITESPACES): Define.
	(PARSE_BASENAME): Likewise.
	(PARSE_INDICANT): Likewise.
	(a68_process_module_map): New function.
	* a68-lang.cc: (a68_init): Move initialization of
	A68_MODULE_FILES from there...
	(a68_init_options): to here.
	(a68_handle_option): Handle OPT_fmodules_map and
	OPT_fmodules_map_.
	* a68-parser-pragmat.cc (handle_access_in_pragmat): Normalize
	module indicants to upper case.
	* ga68.texi (Module search options): New section.