xtensa: Remove variables only used to pass the return of end_sequence()
As a result of the automatic replacement by commit 4dd13988c93c24ba3605f4b9cafc97515c34f2ac,
there are several code fragments that receive the return value of
end_sequence() and immediately use it as the return value of the function
itself.

   rtx_insn *insn;
   ...
   insn = end_sequence ();
   return insn;

It is clear that in such cases, it would be more natural to pass the return
value of end_sequence() directly to the return statement without passing it
through a variable.

Applying this patch naturally does not change any functionality.

gcc/ChangeLog:

	* config/xtensa/xtensa.cc
	(xtensa_expand_block_set_libcall,
	xtensa_expand_block_set_unrolled_loop,
	xtensa_expand_block_set_small_loop, xtensa_call_tls_desc):
	Change the return statement to pass the return value of
	end_sequence() directly without going through a variable, and
	remove the definition of that variable.