forwprop: Check type conversion in pack/unpack [PR123117].
When using pack or unpack in the simplification of a vector constructor
we must make sure that the original BIT_FIELD_REF was no sign-changing
nop conversion.  If it was we cannot safely pack/unpack as that would
skip sign or zero extensions.  This patch adds useless_type_conversion_p
to both paths.

	PR tree-optimization/123117

gcc/ChangeLog:

	* tree-ssa-forwprop.cc (simplify_vector_constructor):
	Check if we had a nop conversion and don't use pack/unpack in
	that case.

gcc/testsuite/ChangeLog:

	* gcc.target/loongarch/vector/lsx/pr123117.c: New test.